Extract necessary review bits.
