+++ app/indexing_services/indexing_service.rb
@@ -6,8 +6,14 @@ class IndexingService < ActiveFedora::IndexingService
 
   def generate_solr_document
     super.tap do |solr_doc|
-      solr_doc.merge!('public_bsi' => object.public?)
-      solr_doc.merge!('reviewed_bsi' => object.reviewed?)
+      solr_doc.merge!('public_bsi' => decorated_object.public?)
+      solr_doc.merge!('reviewed_bsi' => decorated_object.reviewed?)
     end
+  end
+
+  private
+
+  def decorated_object
+    @decorated_object ||= Reviewable.new(object)
   end
 end
