+++ app/models/generic_asset.rb
@@ -9,7 +9,6 @@ class GenericAsset < ActiveFedora::Base
   contains "workflow_metadata", :class_name => "Files::YmlFile"
   delegate :content_changed?,:blank?, :to => :content, :prefix => true
   delegate :streamable_content, :to => :content
-  delegate :reviewed=, :to => :workflow_metadata
 
   def injector
     @injector ||= OregonDigital.inject
@@ -19,22 +18,6 @@ class GenericAsset < ActiveFedora::Base
   # properties.
   def indexing_service
     @indexing_service ||= IndexingService.new(self)
-  end
-
-  def public?
-    read_groups.include?('public')
-  end
-
-  def public=(val)
-    if val == true
-      self.read_groups |= ['public']
-    elsif val == false
-      self.read_groups -= ['public']
-    end
-  end
-
-  def reviewed?
-    !!workflow_metadata.reviewed
   end
 
   private
