+++ spec/models/generic_asset_spec.rb
@@ -144,95 +144,4 @@ RSpec.describe GenericAsset do
     end
   end
 
-  describe "#public?" do
-    context "when it has a public group" do
-      it "should be true" do
-        g = GenericAsset.new
-        g.read_groups = ['public']
-
-        expect(g).to be_public
-      end
-      it "should index it" do
-        g = GenericAsset.new
-        g.read_groups = ['public']
-
-        expect(g.to_solr["public_bsi"]).to eq true
-      end
-    end
-    context "when it doesn't" do
-      it "should be false" do
-        expect(GenericAsset.new).not_to be_public
-      end
-      it "should index it" do
-        g = GenericAsset.new
-        
-        expect(g.to_solr["public_bsi"]).to eq false
-      end
-    end
-  end
-
-  describe "#public=" do
-    context "when set to true" do
-      it "should add public to read groups" do
-        subject.public = true
-
-        expect(subject).to be_public
-      end
-    end
-    context "when set to false" do
-      it "should add public to read groups" do
-        subject.public = true
-        subject.public = false
-
-        expect(subject).not_to be_public
-      end
-    end
-  end
-
-  describe "#reviewed?" do
-    context "when the workflow data says it's reviewed" do
-      it "should be true" do
-        g = GenericAsset.new
-        g.workflow_metadata.reviewed = true
-
-        expect(g).to be_reviewed
-      end
-      it "should index" do
-        g = GenericAsset.new
-        g.workflow_metadata.reviewed = true
-
-        expect(g.to_solr["reviewed_bsi"]).to eq true
-      end
-    end
-    context "when the workflow data doesn't say it's reviewed" do
-      it "should be false" do
-        g = GenericAsset.new
-
-        expect(g).not_to be_reviewed
-      end
-      it "should index" do
-        g = GenericAsset.new
-
-        expect(g.to_solr["reviewed_bsi"]).to eq false
-      end
-    end
-  end
-
-  describe "#reviewed=" do
-    context "when set to true" do
-      it "should mark it was reviewed" do
-        subject.reviewed = true
-
-        expect(subject).to be_reviewed
-      end
-    end
-    context "when set to false" do
-      it "mark it as unreviewed" do
-        subject.reviewed = true
-        subject.reviewed = false
-
-        expect(subject).not_to be_reviewed
-      end
-    end
-  end
 end
